Test cases for register file
This commit introduces a concrete test bench for the register file.
Note that the cpu_defs_package had to be "updated" because otherwise
the simulation failed.
Apart from that, there is a minor bugfix in the register file itself.

Note that our register file only works if we sequentially update the
demux value first (we can do that asynchronously) and only after waiting
a short amount of time, a rising edge leads to a "correct" update of
the output signals.